refactor: moves tests to respective folders
